Add support for WADAM limiting frequencies
If SIF/SIU contains results of omega=0 or -1, the _lim_freq flag will
be set to True. The limiting frequency will only be used for added mass,
but for potential damping, motion RAO, force transfer functions, etc.,
the last two elements (omega = 0 and -1) in wave frequency vector should
be ignored.